refactor: move signup modal to modals
